server/requirements.txt
pillow>=3.0.0
requests
numpy
flask
